+++ src/lib/output/themes/default/partials/header.tsx
@@ -2,19 +2,39 @@ import { classNames, getDisplayName, hasTypeParameters, join } from "../../lib";
 import { JSX } from "../../../../utils";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext";
 import type { PageEvent } from "../../../events";
-import { type Reflection, ReflectionKind } from "../../../../models";
+import type { Reflection } from "../../../../models";
 
 export const header = (context: DefaultThemeRenderContext, props: PageEvent<Reflection>) => {
-    const HeadingLevel = props.model.isProject() ? "h2" : "h1";
+    const opts = context.options.getValue("headings");
+
+    // Don't render on the index page or the class hierarchy page
+    // We should probably someday render on the class hierarchy page, but currently breadcrumbs
+    // are entirely dependent on the reflection hierarchy, so it doesn't make sense today.
+    const renderBreadcrumbs = props.url !== "index.html" && props.url !== "hierarchy.html";
+
+    // Titles are always rendered on DeclarationReflection pages and the modules page for the project.
+    // They are also rendered on the readme + document pages if configured to do so by the user.
+    let renderTitle: boolean;
+    let titleKindString = "";
+    if (props.model.isProject()) {
+        if (props.url === "index.html" && props.model.readme?.length) {
+            renderTitle = opts.readme;
+        } else {
+            renderTitle = true;
+        }
+    } else if (props.model.isDocument()) {
+        renderTitle = opts.document;
+    } else {
+        renderTitle = true;
+        titleKindString = " " + context.internationalization.kindSingularString(props.model.kind);
+    }
+
     return (
         <div class="tsd-page-title">
-            {props.url !== "index.html" && props.url !== "hierarchy.html" && (
-                <ul class="tsd-breadcrumb">{context.breadcrumb(props.model)}</ul>
-            )}
-            {!props.model.isDocument() && (
-                <HeadingLevel class={classNames({ deprecated: props.model.isDeprecated() })}>
-                    {props.model.kind !== ReflectionKind.Project &&
-                        `${context.internationalization.kindSingularString(props.model.kind)} `}
+            {renderBreadcrumbs && <ul class="tsd-breadcrumb">{context.breadcrumb(props.model)}</ul>}
+            {renderTitle && (
+                <h1 class={classNames({ deprecated: props.model.isDeprecated() })}>
+                    {titleKindString}
                     {getDisplayName(props.model)}
                     {hasTypeParameters(props.model) && (
                         <>
@@ -24,7 +44,7 @@ export const header = (context: DefaultThemeRenderContext, props: PageEvent<Refl
                         </>
                     )}
                     {context.reflectionFlags(props.model)}
-                </HeadingLevel>
+                </h1>
             )}
         </div>
     );